+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Netivo CLI Changelog
 
+## Version 1.2.5
+- Fixed blok generation no block name in block.json and wrong style directory
+- Added default options to block.json: icon, keywords
+
 ## Version 1.2.4
 - Fixed pre-commit